Remove Make class, remove unnecessary package
