+++ modules/core/package.json
@@ -19,7 +19,6 @@
     "dotenv": "^16.0.3",
     "ejs": "^3.1.8",
     "express": "^4.18.2",
-    "fs-extra": "^11.1.0",
     "kolorist": "^1.7.0",
     "lodash-es": "^4.17.21",
     "node-fetch": "^3.3.0",
